requirements.txt
langchain-openai>=0.0.5
langchain-core>=0.1.17
python-dotenv>=1.0.0
gradio>=4.14.0
openai>=1.12.0
